% Generated by maven-r-jsr223-plugin: do not edit by hand
\name{fluentSetMessage}
\title{sets a message}
\usage{

}
\arguments{
\item{api}{the java api object}
\item{message}{expects a java.lang.String}
}
\description{

}
\examples{
}
\value{uk.co.terminological.mavenrjsr233plugintest.TestRApi : }
\keyword{java api}
