% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{JAnotherTestRApi}

\alias{JAnotherTestRApi}

\title{}

\usage{\preformatted{
library(testRapi)
api <- JAnotherTestRApi$new();
}}

\description{
	This has no documentation
}

\details{
	This has no documentation
}

\examples{
## -----------------------------------
## Method `JAnotherTestRApi$throwCatchable`
## -----------------------------------

## -----------------------------------
## Method `JAnotherTestRApi$printMessage`
## -----------------------------------

## -----------------------------------
## Method `JAnotherTestRApi$throwRuntime`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Public methods}{
		\itemize{
			\item \href{#method-new}{\code{JAnotherTestRApi$new(message1, message2)}}
			\item \href{#method-throwCatchable}{\code{JAnotherTestRApi$throwCatchable()}}
			\item \href{#method-printMessage}{\code{JAnotherTestRApi$printMessage()}}
			\item \href{#method-throwRuntime}{\code{JAnotherTestRApi$throwRuntime()}}
			\item \code{JAnotherTestRApi$clone()}
			\item \code{JAnotherTestRApi$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the first constructor is used if there are none annotated
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JAnotherTestRApi$new();
  api$new(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a java.lang.String)}
				\item{message2 - will be used for toString}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.mavenrjsr233plugintest.AnotherTestRApi: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwCatchable"></a>}}
	
	\subsection{Method \code{throwCatchable()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JAnotherTestRApi$new();
  api$throwCatchable()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-printMessage"></a>}}
	
	\subsection{Method \code{printMessage()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JAnotherTestRApi$new();
  api$printMessage()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwRuntime"></a>}}
	
	\subsection{Method \code{throwRuntime()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JAnotherTestRApi$new();
  api$throwRuntime()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}