% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{JTestRApi}

\alias{JTestRApi}

\title{}

\usage{\preformatted{
library(testRapi)
api <- JTestRApi$new();
}}

\description{
	A test of the jsr223 templating
}

\details{
	

this is a details comment
}

\examples{
## -----------------------------------
## Method `JTestRApi$doHelloWorld`
## -----------------------------------
  An example
  Spans many lines

## -----------------------------------
## Method `JTestRApi$fluentSetMessage`
## -----------------------------------

## -----------------------------------
## Method `JTestRApi$getMessage`
## -----------------------------------

## -----------------------------------
## Method `JTestRApi$doSum`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Public methods}{
		\itemize{
			\item \href{#method-new}{\code{JTestRApi$new()}}
			\item \href{#method-doHelloWorld}{\code{JTestRApi$doHelloWorld()}}
			\item \href{#method-fluentSetMessage}{\code{JTestRApi$fluentSetMessage(message)}}
			\item \href{#method-getMessage}{\code{JTestRApi$getMessage()}}
			\item \href{#method-doSum}{\code{JTestRApi$doSum(a, b)}}
			\item \code{JTestRApi$clone()}
			\item \code{JTestRApi$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JTestRApi$new();
  api$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.mavenrjsr233plugintest.TestRApi: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doHelloWorld"></a>}}
	
	\subsection{Method \code{doHelloWorld()}}{
		Description of a hello world function
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JTestRApi$new();
  api$doHelloWorld()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: this java method returns a String
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  An example
  Spans many lines
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-fluentSetMessage"></a>}}
	
	\subsection{Method \code{fluentSetMessage()}}{
		A fluent method description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JTestRApi$new();
  api$fluentSetMessage(message)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message the message is a string}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		the JTestRApi itself
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-getMessage"></a>}}
	
	\subsection{Method \code{getMessage()}}{
		message desciption
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JTestRApi$new();
  api$getMessage()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: A message previously set or maybe null
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doSum"></a>}}
	
	\subsection{Method \code{doSum()}}{
		The doSum function description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  api <- JTestRApi$new();
  api$doSum(a, b)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{a the A parameter}{ - (java expects a java.lang.Double)}
				\item{b the B parameter}{ - (java expects a java.lang.Double)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.Double: A+B of course
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}