% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{TestRApi}

\alias{TestRApi}

\title{}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$TestRApi$new(...);
}}

\description{
	A test of the jsr223 templating
}

\details{
	

this is a details comment
}

\examples{
## -----------------------------------
## Method `TestRApi$doHelloWorld`
## -----------------------------------
  An example
  Spans many lines

## -----------------------------------
## Method `TestRApi$fluentSetMessage`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$factoryMethod`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$getMessage`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$doSum`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$TestRApi$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-doHelloWorld}{\code{instance$doHelloWorld()}}
			\item \href{#method-fluentSetMessage}{\code{instance$fluentSetMessage(message)}}
			\item \href{#method-factoryMethod}{\code{instance$factoryMethod(a, b)}}
			\item \href{#method-getMessage}{\code{instance$getMessage()}}
			\item \href{#method-doSum}{\code{instance$doSum(a, b)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$TestRApi$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 TestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doHelloWorld"></a>}}
	
	\subsection{Method \code{doHelloWorld()}}{
		Description of a hello world function
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$doHelloWorld()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: this java method returns a String
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  An example
  Spans many lines
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-fluentSetMessage"></a>}}
	
	\subsection{Method \code{fluentSetMessage()}}{
		A fluent method description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$fluentSetMessage(message)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message the message is a string}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 TestRApi object: nothing as this is a fluent method
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-factoryMethod"></a>}}
	
	\subsection{Method \code{factoryMethod()}}{
		A fluent method description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$factoryMethod(a, b)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{a}{ - (java expects a java.lang.String)}
				\item{b}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 AnotherTestRApi object: nothing as this is a fluent method
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-getMessage"></a>}}
	
	\subsection{Method \code{getMessage()}}{
		message desciption
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$getMessage()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: A message previously set or maybe null
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doSum"></a>}}
	
	\subsection{Method \code{doSum()}}{
		The doSum function description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$doSum(a, b)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{a the A parameter}{ - (java expects a java.lang.Double)}
				\item{b the B parameter}{ - (java expects a java.lang.Double)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.Double: A+B of course
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}